# docs/agent-integration/openapi.yaml
openapi: 3.1.0
info:
  title: Ardent Agent Execution API
  version: 1.0.0
  description: >-
    Minimal hosted API spec for LLM agent tooling (ChatGPT/Grok/MCP wrappers).
servers:
  - url: https://api.ardentresearch.xyz
security:
  - ApiKeyAuth: []
paths:
...
  /feed/recent:
    get:
      summary: Recent public feed activity
      security: []
      parameters:
        - in: query
          name: limit
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 12
      responses:
        '200':
          description: Feed rows
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      type: object
  /wallet:
    get:
      summary: Resolve or provision smart wallet
      parameters:
        - in: query
          name: agent_id
          required: true
          schema:
            type: string
        - in: query
          name: chain
          required: false
          schema:
            type: string
            default: ethereum
      responses:
        '200':
          description: Wallet response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WalletResponse'
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
  /wallet/balance:
    get:
      summary: Get native + ERC-20 token balances for an agent's smart wallet
      parameters:
        - in: query
          name: agent_id
          required: true
          schema:
            type: string
        - in: query
          name: chain
          required: false
          schema:
            type: string
            default: ethereum
      responses:
        '200':
          description: Wallet balance response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WalletBalanceResponse'
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
  /simulate:
    post:
      summary: Simulate execution request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExecutionRequest'
      responses:
        '200':
          description: Simulation result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExecutionResponse'
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
  /execute:
    post:
      summary: Execute transaction (or return payment quote)
      parameters:
        - in: header
          name: X-Payment-Proof
          required: false
          description: JSON string proof header for manual payment flow re-submit.
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExecutionRequest'
      responses:
        '200':
          description: Execution accepted/queued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExecutionResponse'
        '402':
          description: Payment required for manual mode
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentRequiredBody'
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
  /protocols/aave-v3/supply/simulate:
    post:
      summary: Simulate Aave V3 Sepolia supply action
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AaveSupplyRequest'
      responses:
        '200':
          description: Simulation result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExecutionResponse'
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
  /protocols/aave-v3/supply:
    post:
      summary: Execute Aave V3 Sepolia supply action
      parameters:
        - in: header
          name: X-Payment-Proof
          required: false
          description: JSON string proof header for manual payment flow re-submit.
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AaveSupplyRequest'
      responses:
        '200':
          description: Execution accepted/queued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExecutionResponse'
        '402':
          description: Payment required for manual mode
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentRequiredBody'
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
  /protocols/aave-v3/withdraw/simulate:
    post:
      summary: Simulate Aave V3 Sepolia withdraw action
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AaveWithdrawRequest'
      responses:
        '200':
          description: Simulation result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExecutionResponse'
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
  /protocols/aave-v3/withdraw:
    post:
      summary: Execute Aave V3 Sepolia withdraw action
      parameters:
        - in: header
          name: X-Payment-Proof
          required: false
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AaveWithdrawRequest'
      responses:
        '200':
          description: Execution accepted/queued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExecutionResponse'
        '402':
          description: Payment required for manual mode
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentRequiredBody'
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
  /protocols/aave-v3/repay/simulate:
    post:
      summary: Simulate Aave V3 Sepolia repay action
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AaveRepayRequest'
      responses:
        '200':
          description: Simulation result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExecutionResponse'
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
  /protocols/aave-v3/repay:
    post:
      summary: Execute Aave V3 Sepolia repay action
      parameters:
        - in: header
          name: X-Payment-Proof
          required: false
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AaveRepayRequest'
      responses:
        '200':
          description: Execution accepted/queued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExecutionResponse'
        '402':
          description: Payment required for manual mode
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentRequiredBody'
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
  /protocols/aave-v3/borrow/simulate:
    post:
      summary: Simulate Aave V3 Sepolia borrow action
      description: Checks projected health factor using Aave account data and oracle price before simulating Pool.borrow.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AaveBorrowRequest'
      responses:
        '200':
          description: Simulation result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExecutionResponse'
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
  /protocols/aave-v3/borrow:
    post:
      summary: Execute Aave V3 Sepolia borrow action
      description: Checks projected health factor using Aave account data and oracle price before executing Pool.borrow.
      parameters:
        - in: header
          name: X-Payment-Proof
          required: false
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AaveBorrowRequest'
      responses:
        '200':
          description: Execution accepted/queued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExecutionResponse'
        '402':
          description: Payment required for manual mode
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentRequiredBody'
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
  /protocols/aave-v3/position:
    get:
      summary: Read Aave V3 Sepolia account data for an agent wallet
      parameters:
        - in: query
          name: agent_id
          required: true
          schema:
            type: string
        - in: query
          name: chain
          required: false
          schema:
            type: string
            enum: [ethereum]
            default: ethereum
      responses:
        '200':
          description: Aave position response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AavePositionResponse'
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  responses:
    ErrorResponse:
      description: Error
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
  schemas:
...
    WalletBalanceResponse:
      type: object
      properties:
        agent_id:
          type: string
        smart_wallet_address:
          type: string
        chain:
          type: string
        native_balance_wei:
          type: string
          description: Native token balance in wei as a decimal string.
        native_balance_formatted:
          type: string
          description: Native balance in the chain's base unit (e.g. ETH).
        tokens:
          type: array
          items:
            $ref: '#/components/schemas/TokenBalance'
      required:
        - agent_id
        - smart_wallet_address
        - chain
        - native_balance_wei
        - native_balance_formatted
        - tokens
    WalletResponse:
      type: object
      properties:
        agent_id:
          type: string
        smart_wallet_address:
          type: string
        deployed:
          type: boolean
        message:
          type: string

      required:
        - agent_id
        - smart_wallet_address
        - deployed
        - message
...